backend/requirements.txt
# Flask Web框架
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1

# 数据库相关
SQLAlchemy==2.0.44
PyMySQL==1.1.2
python-dotenv==1.1.1

# AI API客户端
openai==2.8.1

# 文件处理库
python-docx==1.2.0
PyPDF2==3.0.1
PyMuPDF==1.26.6

# 依赖项
requests==2.25.1
urllib3==2.5.0
idna==3.3
certifi==2025.10.5
click==8.3.1
Werkzeug==3.1.4
Jinja2==3.1.6
itsdangerous==2.2.0
MarkupSafe==3.0.3
blinker==1.9.0
PyJWT==2.8.0

# 认证相关依赖
bcrypt==4.2.0
email-validator==2.2.0
Flask-Mail==0.9.1

# 语音识别
faster-whisper==1.2.1
